requirements.txt
streamlit
pandas
plotly
openpyxl
xlsxwriter
python-docx
pdfplumber
requests
easyocr
opencv-python-headless
